EDU-1549 Update position of other tasks when pushing or updating only one task
We have to update positions of other tasks if task was added in the middle of its
position was changed.
Unobvious points:

* To calculate the right position of task to post we have to find it
position among pushed tasks only.

* We have to update underlying tasks if task was pushed in the middle. But
we need to filter out task that are already on Stepik. To set correct position
we iterate over pushed task, as their position in the list of pushed tasks differs
from index

* If task was moved we have to update positions for tasks that are between
current and previous positions of moved tasks. New positions are calculated
in the same manner as in previous case